Moved stuff out of AnalyticsParams
