+++ src/main/java/gov/nci/WebAnalytics/AnalyticsRequest.java
@@ -1,8 +1,12 @@
 package gov.nci.WebAnalytics;
 
+import java.io.UnsupportedEncodingException;
 import java.net.URI;
+import java.net.URLDecoder;
+import java.util.ArrayList;
 import java.util.List;
 import org.apache.http.NameValuePair;
+import org.apache.http.message.BasicNameValuePair;
 
 public class AnalyticsRequest {
 
@@ -52,15 +56,15 @@ public class AnalyticsRequest {
 		
 		params = new AnalyticsParams(uri);
 		suites = getSuites(uri);
-		setChannel(getChannel(params.getAll()));
-		channel = getChannel(params.getAll());
-		events = getEvents(params.getAll());
-		props = getProps(params.getAll());
-		eVars = getEvars(params.getAll());
-		hiers = getHiers(params.getAll());
-		linkType = getLinkType(params.getAll());
-		linkName = getLinkName(params.getAll());
-		linkUrl = getLinkUrl(params.getAll());
+		// setChannel(getChannel(params.getAll()));
+		channel = getChannel(buildParamsList(uri));
+		events = getEvents(buildParamsList(uri));
+		props = getProps(buildParamsList(uri));
+		eVars = getEvars(buildParamsList(uri));
+		hiers = getHiers(buildParamsList(uri));
+		linkType = getLinkType(buildParamsList(uri));
+		linkName = getLinkName(buildParamsList(uri));
+		linkUrl = getLinkUrl(buildParamsList(uri));
 	}
 	
 	/**
@@ -132,7 +136,7 @@ public class AnalyticsRequest {
 	 * @return
 	 */
 	public List<NameValuePair> getProps(List<NameValuePair> parms) {
-		return AnalyticsParams.getNumberedParams(parms, AnalyticsParams.PROP_PARTIAL, "prop");
+		return getNumberedParams(parms, AnalyticsParams.PROP_PARTIAL, "prop");
 	}
 	
 	/**
@@ -141,7 +145,7 @@ public class AnalyticsRequest {
 	 * @return
 	 */
 	public List<NameValuePair> getEvars(List<NameValuePair> parms) {
-		return AnalyticsParams.getNumberedParams(parms, AnalyticsParams.EVAR_PARTIAL, "eVar");
+		return getNumberedParams(parms, AnalyticsParams.EVAR_PARTIAL, "eVar");
 	}
 	
 	/**
@@ -150,7 +154,7 @@ public class AnalyticsRequest {
 	 * @return
 	 */
 	public List<NameValuePair> getHiers(List<NameValuePair> parms) {
-		return AnalyticsParams.getNumberedParams(parms, AnalyticsParams.HIER_PARTIAL, "hier");
+		return getNumberedParams(parms, AnalyticsParams.HIER_PARTIAL, "hier");
 	}
 
 	/**
@@ -225,5 +229,102 @@ public class AnalyticsRequest {
 	protected static void nap() {
 		nap(10);
 	}
+	
+
+	
+	
+	
+	/***** URL Param methods *****/
+
+	/**
+	 * Split URI into list of encoded elements
+	 * @param uri
+	 * @return retParams
+	 */
+	public List<NameValuePair> buildParamsList(URI uri) {		
+		List<NameValuePair> rtnParams = new ArrayList<NameValuePair>();
+		
+		try {
+			String queries = getRawQueryString(uri);
+			for(String parm : queries.split("&")) {
+				String[] pair = parm.split("=");
+				String key = URLDecoder.decode(pair[0], "UTF-8");
+				String value = "";
+				if(pair.length > 1) {
+					value = URLDecoder.decode(pair[1], "UTF-8"); 
+				}
+				rtnParams.add(new BasicNameValuePair(key, value));				
+			}
+		} 
+		catch (UnsupportedEncodingException ex) {
+			System.out.println("Error decoding URI in WaParams:buildParamsList()");
+		}		
+		return rtnParams;
+	}
+	
+	/**
+	 * Get an encoded query string from a given URI.
+	 * @param uri
+	 * @return
+	 */	
+	public static String getRawQueryString(URI uri) {
+		return uri.getRawQuery();
+	}
+	
+	/**
+	 * Get the clean query string from a given URI.
+	 * @param uri
+	 * @return
+	 */
+	public static String getQueryString(URI uri) {
+		return uri.getQuery();
+	}
+	
+	/**
+	 * Check query params to see if this is a link tracking event
+	 * @param parms
+	 * @return
+	 */
+	public static boolean hasParam(List<NameValuePair> paramList, String myParam) {
+		for (NameValuePair param : paramList) {
+			if (param.getName().toLowerCase().equals(myParam)) {
+				return true;
+			}
+		}
+		return false;
+	}
+	
+	/**
+	 * Get a list of numbered parameters and their values (e.g. [prop1="www.cancer.gov", prop2="/home", prop3="NCI"])
+	 * @param paramList
+	 * @param parm
+	 * @param replacement
+	 * @return
+	 */
+	public static List<NameValuePair> getNumberedParams(List<NameValuePair> paramList, String parm, String replacement) {
+		List<NameValuePair> rtnList = new ArrayList<>();
+		for (NameValuePair param : paramList) {
+			// Regex: parameter name followed by 1 or more digits, starting with 1-9 only
+			if(param.getName().matches("^" + parm + "[1-9]\\d*$")) {
+				String rtnName = param.getName().replace(parm, replacement);
+				String rtnValue = param.getValue();
+				rtnList.add(new BasicNameValuePair(rtnName, rtnValue));
+			}
+		}
+		return rtnList;
+	}
+	
+	/**
+	 * Overload for getNumberedParams
+	 * @param paramList
+	 * @param parm
+	 * @return
+	 */
+	public static List<NameValuePair> getNumberedParams(List<NameValuePair> paramList, String parm) {
+		return getNumberedParams(paramList, parm, parm);
+	}	
+	
+	
+	
 
 }
